@@ -1,32 +1,32 @@
-name: recipebook
+name: recipe_book_flutter
 description: All recipes in one place
 publish_to: 'none'
-version: 1.0.0+1
-
+version: 1.1.0+2
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: '>=2.19.0 <4.0.0'
+  flutter: '>=3.7.0'
 
 dependencies:
-  beamer: ^0.14.1 # навигация
-  cached_network_image: ^3.1.0
-  cupertino_icons: ^1.0.3
-  dio: ^4.0.0 # взаимодействие с API
-  dotted_border: ^2.0.0+1
-  file_picker: ^4.0.0
-  fish_redux: ^0.3.7
+  beamer: 0.14.1
+  cached_network_image: ^3.2.3
+  cupertino_icons: ^1.0.5
+  dio: ^5.0.1
+  dotted_border: ^2.0.0+3
+  file_picker: ^5.2.5
   flutter:
     sdk: flutter
-  flutter_svg: ^0.22.0
-  google_fonts: ^2.1.0
-  json_annotation: ^4.0.1
-  provider: ^6.0.0
-  url_launcher: ^6.0.9 # запуск внешних ссылок
-  url_strategy: ^0.2.0 # убирает '#' в url
+  flutter_svg: ^2.0.2
+  google_fonts: ^4.0.3
+  json_annotation: ^4.8.0
+  provider: ^6.0.5
+  url_launcher: ^6.1.10
+  url_strategy: ^0.2.0
 
 dev_dependencies:
-  build_runner: ^2.1.1
-  json_serializable: ^4.1.4
-  lint: ^1.6.0
+  build_runner: ^2.3.3
+  carapacik_lints: ^1.3.1
+  json_serializable: ^6.6.1
+  swagger_parser: ^0.8.0
 
 flutter:
   uses-material-design: true
