@@ -1,21 +1,23 @@
+import 'dart:async';
 import 'dart:convert';
 
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
-import 'package:recipebook/notifier/recipe_notifier.dart';
-import 'package:recipebook/resources/images.dart';
-import 'package:recipebook/resources/palette.dart';
-import 'package:recipebook/service/api_service.dart';
-import 'package:recipebook/theme.dart';
-import 'package:recipebook/widgets/header_buttons.dart';
-import 'package:recipebook/widgets/header_widget.dart';
-import 'package:recipebook/widgets/outlined_button.dart';
-import 'package:recipebook/widgets/recipe_list_widget.dart';
+import 'package:recipe_book_flutter/notifier/recipe_notifier.dart';
+import 'package:recipe_book_flutter/resources/images.dart';
+import 'package:recipe_book_flutter/resources/palette.dart';
+import 'package:recipe_book_flutter/service/api_service.dart';
+import 'package:recipe_book_flutter/theme.dart';
+import 'package:recipe_book_flutter/widgets/header_buttons.dart';
+import 'package:recipe_book_flutter/widgets/header_widget.dart';
+import 'package:recipe_book_flutter/widgets/outlined_button.dart';
+import 'package:recipe_book_flutter/widgets/recipe_list_widget.dart';
 
 class FavoritePage extends StatefulWidget {
-  const FavoritePage({Key? key}) : super(key: key);
+  const FavoritePage({super.key});
 
   @override
   _FavoritePageState createState() => _FavoritePageState();
@@ -27,13 +29,18 @@ class _FavoritePageState extends State<FavoritePage> {
   bool isEndOfList = true;
   int skipCounter = 0;
 
-  Future getMoreRecipes() async {
-    Response response;
+  Future<void> getMoreRecipes() async {
+    Response<dynamic> response;
 
     try {
-      response = await _apiService.getRequestWithParam(endPoint: "recipes/favorite", take: 4, skip: skipCounter);
+      response = await _apiService.getRequestWithParam(
+        endPoint: 'recipes/favorite',
+        take: 4,
+        skip: skipCounter,
+      );
       if (response.statusCode == 200) {
-        final listOfRecipes = jsonDecode(response.data as String) as List<dynamic>;
+        final listOfRecipes =
+            jsonDecode(response.data as String) as List<dynamic>;
         if (listOfRecipes.length == 4) {
           setState(() {
             isEndOfList = false;
@@ -45,45 +52,45 @@ class _FavoritePageState extends State<FavoritePage> {
         }
         _recipeNotifier.addRecipes(listOfRecipes);
         skipCounter += 4;
-      } else {
-        // затычка, код не 200
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
-  Future getInitialRecipes() async {
-    Response response;
+  Future<void> getInitialRecipes() async {
+    Response<dynamic> response;
 
     try {
-      response = await _apiService.getInitialWithParam("recipes/favorite", 2);
+      response = await _apiService.getInitialWithParam('recipes/favorite', 2);
       if (response.statusCode == 200) {
-        final listOfRecipes = jsonDecode(response.data as String) as List<dynamic>;
+        final listOfRecipes =
+            jsonDecode(response.data as String) as List<dynamic>;
         if (listOfRecipes.length == 2) {
           setState(() {
             isEndOfList = false;
           });
         }
-        _recipeNotifier.resultString = "Ваш список пуст";
-        _recipeNotifier.addClearRecipes(listOfRecipes);
+        _recipeNotifier
+          ..resultString = 'Ваш список пуст'
+          ..addClearRecipes(listOfRecipes);
         skipCounter += 2;
-      } else {
-        // затычка, код не 200
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
   @override
   void initState() {
+    super.initState();
     _recipeNotifier = Provider.of<RecipeNotifier>(context, listen: false);
     _apiService = ApiService();
-    getInitialRecipes();
-    super.initState();
+    unawaited(getInitialRecipes());
   }
 
   @override
@@ -94,8 +101,9 @@ class _FavoritePageState extends State<FavoritePage> {
           children: [
             SvgPicture.asset(
               CookingImages.wave1,
-              color: Palette.wave,
               width: MediaQuery.of(context).size.width,
+              colorFilter:
+                  const ColorFilter.mode(Palette.wave, BlendMode.srcIn),
             ),
             const HeaderWidget(currentSelectedPage: HeaderButtons.favorite),
             Center(
@@ -107,7 +115,7 @@ class _FavoritePageState extends State<FavoritePage> {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     Text(
-                      "Избранное",
+                      'Избранное',
                       style: Theme.of(context).textTheme.b42,
                     ),
                     const SizedBox(height: 50),
@@ -116,12 +124,10 @@ class _FavoritePageState extends State<FavoritePage> {
                     if (!isEndOfList)
                       Center(
                         child: ButtonOutlinedWidget(
-                          text: "Загрузить еще",
+                          text: 'Загрузить еще',
                           width: 309,
                           height: 60,
-                          onPressed: () {
-                            getMoreRecipes();
-                          },
+                          onPressed: getMoreRecipes,
                         ),
                       ),
                     const SizedBox(height: 100),
