@@ -1,12 +1,10 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:recipebook/resources/palette.dart';
+import 'package:recipe_book_flutter/resources/palette.dart';
 
 final themeData = ThemeData.light().copyWith(
   primaryColor: Palette.orange,
   scaffoldBackgroundColor: Colors.white,
-  backgroundColor: Colors.white,
   iconTheme: const IconThemeData(color: Palette.orange),
   primaryTextTheme: GoogleFonts.montserratTextTheme(),
   textTheme: _textTheme,
@@ -15,70 +13,70 @@ final themeData = ThemeData.light().copyWith(
 TextTheme get _textTheme => GoogleFonts.montserratTextTheme().copyWith();
 
 extension TextThemeExt on TextTheme {
-  TextStyle get b107 => headline2!.copyWith(
+  TextStyle get b107 => displayMedium!.copyWith(
         color: Palette.orange,
         fontSize: 107,
         fontWeight: FontWeight.w600,
       );
 
-  TextStyle get b72 => headline1!.copyWith(
+  TextStyle get b72 => displayLarge!.copyWith(
         color: Palette.main,
         fontSize: 72,
         fontWeight: FontWeight.w700,
         height: 87 / 72,
       );
 
-  TextStyle get b42 => headline2!.copyWith(
+  TextStyle get b42 => displayMedium!.copyWith(
         color: Palette.main,
         fontSize: 42,
         fontWeight: FontWeight.w700,
         height: 62 / 42,
       );
 
-  TextStyle get b24 => headline3!.copyWith(
+  TextStyle get b24 => displaySmall!.copyWith(
         color: Palette.main,
         fontSize: 24,
         fontWeight: FontWeight.w700,
       );
 
-  TextStyle get m24 => headline3!.copyWith(
+  TextStyle get m24 => displaySmall!.copyWith(
         color: Palette.orange.withOpacity(0.5),
         fontSize: 24,
         fontWeight: FontWeight.w600,
       );
 
-  TextStyle get b20 => headline4!.copyWith(
+  TextStyle get b20 => headlineMedium!.copyWith(
         color: Palette.mainLighten1,
         fontSize: 20,
         fontWeight: FontWeight.w700,
         height: 30 / 20,
       );
 
-  TextStyle get b18 => bodyText1!.copyWith(
+  TextStyle get b18 => bodyLarge!.copyWith(
         color: Palette.main,
         fontSize: 18,
         fontWeight: FontWeight.w600,
       );
 
-  TextStyle get r18 => bodyText1!.copyWith(
+  TextStyle get r18 => bodyLarge!.copyWith(
         color: Palette.mainLighten1,
         fontSize: 18,
         fontWeight: FontWeight.w400,
       );
 
-  TextStyle get r16 => bodyText2!.copyWith(
+  TextStyle get r16 => bodyMedium!.copyWith(
         color: Palette.main.withOpacity(0.5),
         fontSize: 16,
         fontWeight: FontWeight.w400,
       );
 
-  TextStyle get r14 => bodyText2!.copyWith(
+  TextStyle get r14 => bodyMedium!.copyWith(
         color: Palette.main.withOpacity(0.5),
         fontSize: 14,
         fontWeight: FontWeight.w400,
       );
 
-  TextStyle get n18 => bodyText2!.copyWith(
+  TextStyle get n18 => bodyMedium!.copyWith(
         color: Palette.orange,
         fontSize: 18,
         fontWeight: FontWeight.w500,
