@@ -4,15 +4,16 @@ part 'auth_result.g.dart';
 
 @JsonSerializable()
 class AuthResult {
-  AuthResult({
+  const AuthResult({
     required this.isSuccess,
     required this.errorMessage,
   });
 
+  factory AuthResult.fromJson(Map<String, dynamic> json) =>
+      _$AuthResultFromJson(json);
+
   final bool isSuccess;
   final String? errorMessage;
-
-  factory AuthResult.fromJson(Map<String, dynamic> json) => _$AuthResultFromJson(json);
 
   Map<String, dynamic> toJson() => _$AuthResultToJson(this);
 }
