@@ -6,14 +6,15 @@ part of 'ingredient.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Ingredient _$IngredientFromJson(Map<String, dynamic> json) {
-  return Ingredient(
-    title: json['title'] as String,
-    ingredientNames: (json['ingredientNames'] as List<dynamic>).map((e) => e as String).toList(),
-  );
-}
+Ingredient _$IngredientFromJson(Map<String, dynamic> json) => Ingredient(
+      title: json['title'] as String,
+      ingredientNames: (json['ingredientNames'] as List<dynamic>)
+          .map((e) => e as String)
+          .toList(),
+    );
 
-Map<String, dynamic> _$IngredientToJson(Ingredient instance) => <String, dynamic>{
+Map<String, dynamic> _$IngredientToJson(Ingredient instance) =>
+    <String, dynamic>{
       'title': instance.title,
       'ingredientNames': instance.ingredientNames,
     };
