@@ -6,17 +6,18 @@ part 'profile_command.g.dart';
 class ProfileCommand {
   ProfileCommand({
     required this.name,
-    this.description,
     required this.login,
+    this.description,
     this.password,
   });
+
+  factory ProfileCommand.fromJson(Map<String, dynamic> json) =>
+      _$ProfileCommandFromJson(json);
 
   String name;
   String? description;
   String login;
   String? password;
-
-  factory ProfileCommand.fromJson(Map<String, dynamic> json) => _$ProfileCommandFromJson(json);
 
   Map<String, dynamic> toJson() => _$ProfileCommandToJson(this);
 }
