@@ -19,6 +19,9 @@ class RecipeItem {
     required this.tags,
   });
 
+  factory RecipeItem.fromJson(Map<String, dynamic> json) =>
+      _$RecipeItemFromJson(json);
+
   final int recipeId;
   final String title;
   final String description;
@@ -31,8 +34,6 @@ class RecipeItem {
   bool isLiked;
   bool isFavorite;
   final List<String> tags;
-
-  factory RecipeItem.fromJson(Map<String, dynamic> json) => _$RecipeItemFromJson(json);
 
   Map<String, dynamic> toJson() => _$RecipeItemToJson(this);
 }
