@@ -6,12 +6,10 @@ part of 'user_model.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-UserModel _$UserModelFromJson(Map<String, dynamic> json) {
-  return UserModel(
-    name: json['name'] as String,
-    login: json['login'] as String,
-  );
-}
+UserModel _$UserModelFromJson(Map<String, dynamic> json) => UserModel(
+      name: json['name'] as String,
+      login: json['login'] as String,
+    );
 
 Map<String, dynamic> _$UserModelToJson(UserModel instance) => <String, dynamic>{
       'name': instance.name,
