@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-import 'package:flutter/material.dart';
-import 'package:recipebook/model/user_model.dart';
-import 'package:recipebook/service/api_service.dart';
+import 'package:flutter/foundation.dart';
+import 'package:recipe_book_flutter/model/user_model.dart';
+import 'package:recipe_book_flutter/service/api_service.dart';
 
 class AuthNotifier extends ChangeNotifier {
   UserModel? _userDetail;
@@ -10,41 +10,41 @@ class AuthNotifier extends ChangeNotifier {
 
   UserModel? get userDetail => _userDetail;
 
-  void set userDetail(UserModel? userDetail) {
+  set userDetail(UserModel? userDetail) {
     _userDetail = userDetail;
     notifyListeners();
   }
 
   bool get isAuth => _userDetail != null;
 
-  Future getCurrentUser() async {
+  Future<void> getCurrentUser() async {
     try {
-      final response = await _apiService.getRequest("user/current-user");
+      final response = await _apiService.getRequest('user/current-user');
       if (response.statusCode == 200) {
-        final user = UserModel.fromJson(jsonDecode(response.data as String) as Map<String, dynamic>);
+        final user = UserModel.fromJson(
+          jsonDecode(response.data as String) as Map<String, dynamic>,
+        );
         userDetail = user;
       } else if (response.statusCode == 403) {
         userDetail = null;
-      } else {
-        //затычка
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
-  Future logout() async {
+  Future<void> logout() async {
     try {
-      final response = await _apiService.postRequestWithoutData("/user/logout");
+      final response = await _apiService.postRequestWithoutData('/user/logout');
       if (response.statusCode == 200) {
         userDetail = null;
-      } else {
-        //
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 }
