@@ -1,24 +1,27 @@
 import 'package:flutter/material.dart';
-import 'package:recipebook/model/ingredient.dart';
+import 'package:recipe_book_flutter/model/ingredient.dart';
 
 class IngredientNotifier extends ChangeNotifier {
-  List<Ingredient> _ingredientList = [Ingredient(title: "", ingredientNames: [])];
+  List<Ingredient> _ingredientList = [
+    Ingredient(title: '', ingredientNames: [])
+  ];
 
   List<Ingredient> get ingredientList => _ingredientList;
 
-  replaceList(List<Ingredient> newList) => _ingredientList = newList;
+  List<Ingredient> replaceList(List<Ingredient> newList) =>
+      _ingredientList = newList;
 
-  addNewIngredient() {
-    _ingredientList.add(Ingredient(title: "", ingredientNames: []));
+  void addNewIngredient() {
+    _ingredientList.add(Ingredient(title: '', ingredientNames: []));
     notifyListeners();
   }
 
-  deleteIngredient(int index) {
+  void deleteIngredient(int index) {
     _ingredientList.removeAt(index);
     notifyListeners();
   }
 
-  clearList(){
-    _ingredientList = [Ingredient(title: "", ingredientNames: [])];
+  void clearList() {
+    _ingredientList = [Ingredient(title: '', ingredientNames: [])];
   }
 }
