@@ -1,27 +1,28 @@
 import 'dart:collection';
 
 import 'package:flutter/material.dart';
-import 'package:recipebook/model/recipe_item.dart';
+import 'package:recipe_book_flutter/model/recipe_item.dart';
 
 class RecipeNotifier extends ChangeNotifier {
-  List<RecipeItem> _recipeList = [];
-  String resultString = "Идёт загрузка...";
+  final List<RecipeItem> _recipeList = [];
+  String resultString = 'Идёт загрузка...';
 
-  UnmodifiableListView<RecipeItem> get recipeList => UnmodifiableListView(_recipeList);
+  UnmodifiableListView<RecipeItem> get recipeList =>
+      UnmodifiableListView(_recipeList);
 
-  addClearRecipes(List<dynamic> data) {
+  void addClearRecipes(List<dynamic> data) {
     _recipeList.clear();
     addRecipes(data);
   }
 
-  addRecipes(List<dynamic> data) {
+  void addRecipes(List<dynamic> data) {
     for (final item in data) {
       _recipeList.add(RecipeItem.fromJson(item as Map<String, dynamic>));
     }
     notifyListeners();
   }
 
-  removeItem(int recipeId) {
+  void removeItem(int recipeId) {
     _recipeList.removeWhere((item) => item.recipeId == recipeId);
     notifyListeners();
   }
