@@ -1,23 +1,23 @@
 import 'package:flutter/material.dart';
 
 class StepNotifier extends ChangeNotifier {
-  List<String> _stepList = [""];
+  List<String> _stepList = [''];
 
   List<String> get stepList => _stepList;
 
-  replaceList(List<String> newList) => _stepList = newList;
+  List<String> replaceList(List<String> newList) => _stepList = newList;
 
-  addStep() {
-    _stepList.add("");
+  void addStep() {
+    _stepList.add('');
     notifyListeners();
   }
 
-  deleteStep(int index) {
+  void deleteStep(int index) {
     _stepList.removeAt(index);
     notifyListeners();
   }
 
-  clearList(){
-    _stepList = [""];
+  void clearList() {
+    _stepList = [''];
   }
 }
