@@ -1,12 +1,12 @@
-class CookingIcons {
-  CookingIcons._();
+abstract class CookingIcons {
+  const CookingIcons._();
 
-  static String chef = "assets/icons/chef.svg";
-  static String confetti = "assets/icons/confetti.svg";
-  static String cook = "assets/icons/cook.svg";
-  static String exit = "assets/icons/exit.svg";
-  static String login = "assets/icons/login.svg";
-  static String menu = "assets/icons/menu.svg";
-  static String recipeOfDay = "assets/icons/recipe_of_day.svg";
-  static String upload = "assets/icons/upload.svg";
+  static String chef = 'assets/icons/ic_chef.svg';
+  static String confetti = 'assets/icons/ic_confetti.svg';
+  static String cook = 'assets/icons/ic_cook.svg';
+  static String exit = 'assets/icons/ic_exit.svg';
+  static String login = 'assets/icons/ic_login.svg';
+  static String menu = 'assets/icons/ic_menu.svg';
+  static String recipeOfDay = 'assets/icons/ic_recipe_of_day.svg';
+  static String upload = 'assets/icons/ic_upload.svg';
 }
