@@ -1,7 +1,7 @@
-class CookingImages {
-  CookingImages._();
+abstract class CookingImages {
+  const CookingImages._();
 
-  static String homeBackground = "assets/images/home_background.png";
-  static String wave1 = "assets/images/wave1.svg";
-  static String wave2 = "assets/images/wave2.svg";
+  static String homeBackground = 'assets/images/home_background.png';
+  static String wave1 = 'assets/images/wave1.svg';
+  static String wave2 = 'assets/images/wave2.svg';
 }
