@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:recipebook/resources/palette.dart';
-import 'package:recipebook/theme.dart';
+import 'package:recipe_book_flutter/resources/palette.dart';
+import 'package:recipe_book_flutter/theme.dart';
 
 class RecipeTagsList extends StatelessWidget {
-  const RecipeTagsList({
-    Key? key,
-    required this.tags,
-  }) : super(key: key);
+  const RecipeTagsList({required this.tags, super.key});
 
   final List<String> tags;
 
@@ -20,20 +17,24 @@ class RecipeTagsList extends StatelessWidget {
             shrinkWrap: true,
             scrollDirection: Axis.horizontal,
             itemCount: tags.length > 3 ? 3 : tags.length,
-            itemBuilder: (BuildContext context, int index) {
+            itemBuilder: (context, index) {
               return Container(
-                padding: const EdgeInsets.symmetric(vertical: 2, horizontal: 10),
+                padding:
+                    const EdgeInsets.symmetric(vertical: 2, horizontal: 10),
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(5),
                   color: Palette.orange.withOpacity(0.2),
                 ),
                 child: Text(
                   tags[index].toLowerCase(),
-                  style: Theme.of(context).textTheme.r14.copyWith(color: Palette.orange),
+                  style: Theme.of(context)
+                      .textTheme
+                      .r14
+                      .copyWith(color: Palette.orange),
                 ),
               );
             },
-            separatorBuilder: (BuildContext context, int index) {
+            separatorBuilder: (context, index) {
               return const SizedBox(width: 5);
             },
           ),
