@@ -1,30 +1,33 @@
+import 'dart:async';
 import 'dart:convert';
 
 import 'package:beamer/beamer.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
-import 'package:recipebook/notifier/auth_notifier.dart';
-import 'package:recipebook/notifier/recipe_notifier.dart';
-import 'package:recipebook/resources/icons.dart';
-import 'package:recipebook/resources/images.dart';
-import 'package:recipebook/resources/palette.dart';
-import 'package:recipebook/service/api_service.dart';
-import 'package:recipebook/theme.dart';
-import 'package:recipebook/widgets/category_card.dart';
-import 'package:recipebook/widgets/contained_button.dart';
-import 'package:recipebook/widgets/header_buttons.dart';
-import 'package:recipebook/widgets/header_widget.dart';
-import 'package:recipebook/widgets/not_auth_dialog.dart';
-import 'package:recipebook/widgets/outlined_button.dart';
-import 'package:recipebook/widgets/recipe_list_widget.dart';
+import 'package:recipe_book_flutter/notifier/auth_notifier.dart';
+import 'package:recipe_book_flutter/notifier/recipe_notifier.dart';
+import 'package:recipe_book_flutter/resources/icons.dart';
+import 'package:recipe_book_flutter/resources/images.dart';
+import 'package:recipe_book_flutter/resources/palette.dart';
+import 'package:recipe_book_flutter/service/api_service.dart';
+import 'package:recipe_book_flutter/theme.dart';
+import 'package:recipe_book_flutter/widgets/category_card.dart';
+import 'package:recipe_book_flutter/widgets/contained_button.dart';
+import 'package:recipe_book_flutter/widgets/header_buttons.dart';
+import 'package:recipe_book_flutter/widgets/header_widget.dart';
+import 'package:recipe_book_flutter/widgets/not_auth_dialog.dart';
+import 'package:recipe_book_flutter/widgets/outlined_button.dart';
+import 'package:recipe_book_flutter/widgets/recipe_list_widget.dart';
 
+// ignore: must_be_immutable
 class RecipesPage extends StatefulWidget {
   RecipesPage({
     this.searchQuery,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   String? searchQuery;
 
@@ -39,27 +42,29 @@ class _RecipesPageState extends State<RecipesPage> {
   bool isEndOfList = true;
   int skipCounter = 0;
 
-  Future searchRecipes() async {
-    Response response;
+  Future<void> searchRecipes() async {
+    Response<dynamic> response;
 
     try {
       response = await apiService.getRequestWithParam(
-        endPoint: "recipes",
+        endPoint: 'recipes',
         take: 4,
         skip: skipCounter,
         searchQuery: widget.searchQuery,
       );
       if (response.statusCode == 200) {
-        final listOfRecipes = jsonDecode(response.data as String) as List<dynamic>;
+        final listOfRecipes =
+            jsonDecode(response.data as String) as List<dynamic>;
         if (listOfRecipes.length == 4) {
           setState(() {
             isEndOfList = false;
           });
         }
         if (widget.searchQuery!.isEmpty) {
-          recipeNotifier.resultString = "Ничего не найдено";
+          recipeNotifier.resultString = 'Ничего не найдено';
         } else {
-          recipeNotifier.resultString = """По запросу "${widget.searchQuery}" ничего не найдено""";
+          recipeNotifier.resultString =
+              '''По запросу "${widget.searchQuery}" ничего не найдено''';
         }
         if (skipCounter == 0) {
           recipeNotifier.addClearRecipes(listOfRecipes);
@@ -68,22 +73,26 @@ class _RecipesPageState extends State<RecipesPage> {
         }
         textController?.text = widget.searchQuery!;
         skipCounter += 4;
-      } else {
-        // затычка, код не 200
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
-  Future getMoreRecipes() async {
-    Response response;
+  Future<void> getMoreRecipes() async {
+    Response<dynamic> response;
 
     try {
-      response = await apiService.getRequestWithParam(endPoint: "recipes", take: 4, skip: skipCounter);
+      response = await apiService.getRequestWithParam(
+        endPoint: 'recipes',
+        take: 4,
+        skip: skipCounter,
+      );
       if (response.statusCode == 200) {
-        final listOfRecipes = jsonDecode(response.data as String) as List<dynamic>;
+        final listOfRecipes =
+            jsonDecode(response.data as String) as List<dynamic>;
         if (listOfRecipes.length == 4) {
           setState(() {
             isEndOfList = false;
@@ -95,22 +104,22 @@ class _RecipesPageState extends State<RecipesPage> {
         }
         recipeNotifier.addRecipes(listOfRecipes);
         skipCounter += 4;
-      } else {
-        // затычка, код не 200
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
-  Future getInitialRecipes() async {
-    Response response;
+  Future<void> getInitialRecipes() async {
+    Response<dynamic> response;
 
     try {
-      response = await apiService.getInitialWithParam("recipes", 4);
+      response = await apiService.getInitialWithParam('recipes', 4);
       if (response.statusCode == 200) {
-        final listOfRecipes = jsonDecode(response.data as String) as List<dynamic>;
+        final listOfRecipes =
+            jsonDecode(response.data as String) as List<dynamic>;
         if (listOfRecipes.length == 4) {
           setState(() {
             isEndOfList = false;
@@ -118,12 +127,11 @@ class _RecipesPageState extends State<RecipesPage> {
         }
         recipeNotifier.addClearRecipes(listOfRecipes);
         skipCounter += 4;
-      } else {
-        // затычка, код не 200
-      }
+      } else {}
     } on Exception catch (e) {
-      // возможно перенаправление на отдельную страницу
-      print(e);
+      if (kDebugMode) {
+        print(e);
+      }
     }
   }
 
@@ -131,9 +139,9 @@ class _RecipesPageState extends State<RecipesPage> {
   void initState() {
     apiService = ApiService();
     if (widget.searchQuery != null) {
-      searchRecipes();
+      unawaited(searchRecipes());
     } else {
-      getInitialRecipes();
+      unawaited(getInitialRecipes());
     }
     super.initState();
   }
@@ -147,15 +155,17 @@ class _RecipesPageState extends State<RecipesPage> {
           children: [
             SvgPicture.asset(
               CookingImages.wave1,
-              color: Palette.wave,
+              colorFilter:
+                  const ColorFilter.mode(Palette.wave, BlendMode.srcIn),
               width: MediaQuery.of(context).size.width,
             ),
             Padding(
               padding: const EdgeInsets.only(top: 900),
               child: SvgPicture.asset(
                 CookingImages.wave2,
-                color: Palette.wave,
                 width: MediaQuery.of(context).size.width,
+                colorFilter:
+                    const ColorFilter.mode(Palette.wave, BlendMode.srcIn),
               ),
             ),
             const HeaderWidget(currentSelectedPage: HeaderButtons.recipes),
@@ -170,22 +180,26 @@ class _RecipesPageState extends State<RecipesPage> {
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
                         Text(
-                          "Рецепты",
+                          'Рецепты',
                           style: Theme.of(context).textTheme.b42,
                         ),
                         Consumer<AuthNotifier>(
-                          builder: (context, auth, child) => ButtonContainedWidget(
+                          builder: (context, auth, child) =>
+                              ButtonContainedWidget(
                             icon: Icons.add,
                             padding: 18,
-                            text: "Добавить рецепт",
+                            text: 'Добавить рецепт',
                             width: 278,
                             height: 60,
                             onPressed: auth.isAuth
                                 ? () {
-                                    context.beamToNamed("/recipes/add");
+                                    context.beamToNamed('/recipes/add');
                                   }
                                 : () {
-                                    notAuthDialog(context, "Добавлять рецепты могут только зарегистрированные пользователи.");
+                                    notAuthDialog(
+                                      context,
+                                      'Добавлять рецепты могут только зарегистрированные пользователи.',
+                                    );
                                   },
                           ),
                         ),
@@ -197,66 +211,74 @@ class _RecipesPageState extends State<RecipesPage> {
                       children: [
                         CategoryCardWidget(
                           iconPath: CookingIcons.menu,
-                          title: "Простые блюда",
-                          onPressed: () {
+                          title: 'Простые блюда',
+                          onPressed: () async {
                             skipCounter = 0;
                             isEndOfList = false;
-                            widget.searchQuery = "простое";
+                            widget.searchQuery = 'простое';
                             context.currentBeamLocation.update(
                               (state) => state.copyWith(
                                 pathBlueprintSegments: ['recipes'],
-                                queryParameters: {'searchQuery': widget.searchQuery!},
+                                queryParameters: {
+                                  'searchQuery': widget.searchQuery!
+                                },
                               ),
                             );
-                            searchRecipes();
+                            await searchRecipes();
                           },
                         ),
                         CategoryCardWidget(
                           iconPath: CookingIcons.cook,
-                          title: "Детское",
-                          onPressed: () {
+                          title: 'Детское',
+                          onPressed: () async {
                             skipCounter = 0;
                             isEndOfList = false;
-                            widget.searchQuery = "детское";
+                            widget.searchQuery = 'детское';
                             context.currentBeamLocation.update(
                               (state) => state.copyWith(
                                 pathBlueprintSegments: ['recipes'],
-                                queryParameters: {'searchQuery': widget.searchQuery!},
+                                queryParameters: {
+                                  'searchQuery': widget.searchQuery!
+                                },
                               ),
                             );
-                            searchRecipes();
+                            await searchRecipes();
                           },
                         ),
                         CategoryCardWidget(
                           iconPath: CookingIcons.chef,
-                          title: "От шеф-поваров",
-                          onPressed: () {
+                          title: 'От шеф-поваров',
+                          onPressed: () async {
                             skipCounter = 0;
                             isEndOfList = false;
-                            widget.searchQuery = "шеф-повар";
+                            widget.searchQuery = 'шеф-повар';
                             context.currentBeamLocation.update(
                               (state) => state.copyWith(
                                 pathBlueprintSegments: ['recipes'],
-                                queryParameters: {'searchQuery': widget.searchQuery!},
+                                queryParameters: {
+                                  'searchQuery': widget.searchQuery!
+                                },
                               ),
                             );
-                            searchRecipes();
+                            await searchRecipes();
                           },
                         ),
                         CategoryCardWidget(
                           iconPath: CookingIcons.confetti,
-                          title: "На праздник",
-                          onPressed: () {
+                          title: 'На праздник',
+                          onPressed: () async {
                             skipCounter = 0;
                             isEndOfList = false;
-                            widget.searchQuery = "праздник";
+                            widget.searchQuery = 'праздник';
                             context.currentBeamLocation.update(
                               (state) => state.copyWith(
                                 pathBlueprintSegments: ['recipes'],
-                                queryParameters: {'searchQuery': widget.searchQuery!},
+                                queryParameters: {
+                                  'searchQuery': widget.searchQuery!
+                                },
                               ),
                             );
-                            searchRecipes();
+                            await searchRecipes();
                           },
                         ),
                       ],
@@ -266,7 +288,7 @@ class _RecipesPageState extends State<RecipesPage> {
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
                         Text(
-                          "Поиск рецепта",
+                          'Поиск рецепта',
                           style: Theme.of(context).textTheme.b24,
                         ),
                         Row(
@@ -275,7 +297,10 @@ class _RecipesPageState extends State<RecipesPage> {
                             Container(
                               height: 73,
                               width: 779,
-                              padding: const EdgeInsets.symmetric(vertical: 28, horizontal: 32),
+                              padding: const EdgeInsets.symmetric(
+                                vertical: 28,
+                                horizontal: 32,
+                              ),
                               decoration: BoxDecoration(
                                 color: Colors.white,
                                 borderRadius: BorderRadius.circular(16),
@@ -290,30 +315,35 @@ class _RecipesPageState extends State<RecipesPage> {
                               child: TextField(
                                 controller: textController,
                                 cursorColor: Palette.orange,
-                                style: Theme.of(context).textTheme.r18.copyWith(color: Palette.main),
+                                style: Theme.of(context)
+                                    .textTheme
+                                    .r18
+                                    .copyWith(color: Palette.main),
                                 decoration: InputDecoration(
                                   border: InputBorder.none,
-                                  hintText: "Название блюда...",
+                                  hintText: 'Название блюда...',
                                   hintStyle: Theme.of(context).textTheme.r16,
                                 ),
                               ),
                             ),
                             const SizedBox(width: 16),
                             ButtonContainedWidget(
-                              text: "Поиск",
+                              text: 'Поиск',
                               width: 152,
                               height: 73,
-                              onPressed: () {
+                              onPressed: () async {
                                 skipCounter = 0;
                                 isEndOfList = false;
                                 widget.searchQuery = textController!.text;
                                 context.currentBeamLocation.update(
                                   (state) => state.copyWith(
                                     pathBlueprintSegments: ['recipes'],
-                                    queryParameters: {'searchQuery': widget.searchQuery!},
+                                    queryParameters: {
+                                      'searchQuery': widget.searchQuery!
+                                    },
                                   ),
                                 );
-                                searchRecipes();
+                                await searchRecipes();
                               },
                             ),
                           ],
@@ -325,12 +355,10 @@ class _RecipesPageState extends State<RecipesPage> {
                     const SizedBox(height: 65),
                     if (!isEndOfList)
                       ButtonOutlinedWidget(
-                        text: "Загрузить еще",
+                        text: 'Загрузить еще',
                         width: 309,
                         height: 60,
-                        onPressed: () {
-                          getMoreRecipes();
-                        },
+                        onPressed: getMoreRecipes,
                       ),
                     const SizedBox(height: 108),
                   ],
